Move PeerKind and ConnectionKind enums to constants and peerType -> peerKind
